+++ src/cores/aws/s3-service.spec.ts
