+++ src/cores/aws/sns-service.spec.ts
