+++ src/cores/core-services.ts
@@ -11,6 +11,15 @@
 import { NextMode, NextContext } from './core-types';
 
 /** ********************************************************************************************************************
+ *  Core Services
+ ** ********************************************************************************************************************/
+export interface CoreKmsService {
+    hello: () => { hello: string };
+    encrypt: (message: string, keyId?: string) => Promise<string>;
+    decrypt: (encryptedSecret: string) => Promise<string>;
+}
+
+/** ********************************************************************************************************************
  *  Protocol Services
  ** ********************************************************************************************************************/
 export type STAGE = 'local' | 'dev' | 'prod';
